refactoring: moved generic functions into `common` module
